Implement Zstd streaming decompression, Refactor res api, Implement Index Allocator, Improve math for avx2, more minor fixes and changes

+++ libraries/lib_awn_win32/include/awn/ukern/ukern_fiberlocalstorage.h
@@ -17,8 +17,8 @@
 
 namespace awn::ukern {
 
-    typedef u32 UKernHandle;
-    typedef u64 UKernCoreMask;
+    typedef u32    UKernHandle;
+    typedef size_t UKernCoreMask;
 
     using ThreadFunction = void (*)(void *);
